Implement one-click Discord RPC bridge
Nero Prefix Settings now has a dependency on QuaZip (which I hope I set it right so the correct version is pulled per Qt5 or Qt6).
For now, wget is polled as an external process dependency to actually grab the bridge zip, because I can't be assed to implement a proper downloader at the moment. Hope to fix this soon.

Also resolves DLL lists writing; since "empty" lists are still lists, trying to write an empty list to the config results in an "@Invalid" string that breaks the list parser in the code. Now we explicitly write a blank string to make the setting null as it should if the DLLs queue is actually empty; the check for if overrides exist in the current config (prefix or shortcut) now more simply checks if it's empty, not look for a first entry that may not exist.