Use "Default Settings" for Windows Version, until explicitly set by user
Originally, every shortcut started with a default setting of "Windows 10", which matches Wine's default. Not only would this get out of date eventually, but it could be confusing for the user giving the impression that every app starts with an override, and was just redundant af.
NeroFS will now set that field blank for new shortcuts, and the Prefix Settings dialog will check to ensure that the field is blank or not (because why doesn't settings.isNull() NOT WORK PROPERLY???) - if so, set the default to the placeholder for WinVer.