Use curl instead of wget
Curl is at least a required component on Arch, and otherwise does the same thing so it's less of a chance that the user won't have this in some form. It also seems to be marginally faster to download the package?
Again, a built-in downloader would be preferable, but this is still enough to work until then.